implement pinned repositories
